Add setup.py and restructure package for distribution
Introduced a setup.py file to the project to facilitate package distribution. The Fortress module was moved into a new package named 'python_fortress' to comply with Python's packaging conventions. Test and documentation files were updated accordingly to reflect these changes. This restructuring makes it easier for potential users to integrate the module into their own projects and simplifies updates in the future.
