+++ python_fortress/fortress.py
